requirements.txt
Django==5.1.3
django-ninja==1.3.0
pillow==11.0.0
PyJWT==2.10.0
email_validator==2.2.0
python-magic==0.4.27
mutagen==1.47.0
